+++ lib/src/iso/8601/encodeIso8601Duration.ts
@@ -1,0 +1,32 @@
+/**
+ * Encode a duration in seconds to an ISO 8601 duration string.
+ *
+ * @param duration - Duration in seconds.
+ * @returns ISO 8601 duration string.
+ *
+ * @group ISO 8601
+ * @beta
+ *
+ * @example
+ * ```ts
+ * const result = encodeIso8601Duration(3661);
+ * console.log(result);
+ * // -> 'PT1H1M1S'
+ * ```
+ */
+export function encodeIso8601Duration(duration: number): string {
+	if (!isFinite(duration)) {
+		return 'PT';
+	}
+
+	const hours = Math.floor(duration / 3600);
+	const minutes = Math.floor((duration % 3600) / 60);
+	const seconds = duration % 60;
+	if (hours > 0) {
+		return `PT${hours}H${minutes}M${seconds}S`;
+	}
+	else if (minutes > 0) {
+		return `PT${minutes}M${seconds}S`;
+	}
+	return `PT${seconds}S`;
+}